made publish reusable
